requirements.txt
asgiref==3.6.0
dj-database-url==2.1.0
Django==4.1.7
django-jazzmin==2.6.0
django-recaptcha==3.0.0
mysql==0.0.3
mysqlclient==2.2.0
python-decouple==3.8
sqlparse==0.4.3
typing_extensions==4.8.0
tzdata==2022.7
